requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.6.0
anthropic==0.55.0
anyio==4.10.0
appnope==0.1.3
asttokens==2.4.1
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.2.1
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.8.0
docker==7.1.0
docstring_parser==0.17.0
elevenlabs==0.2.26
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.31.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.14.0
hf-xet==1.1.8
httpcore==1.0.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.4
importlib_metadata==8.7.0
iniconfig==2.1.0
instructor==1.10.0
invoke==2.2.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-anthropic==0.3.3
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-mistralai==0.2.11
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langgraph==0.2.76
langgraph-checkpoint==2.1.1
langgraph-sdk==0.1.74
langsmith==0.3.45
litellm==1.75.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mcp==1.13.0
mdurl==0.1.2
mistralai==1.9.7
multidict==6.6.4
numpy==1.26.1
openai==1.98.0
opencv-python==4.8.1.78
openpyxl==3.1.5
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
pluggy==1.6.0
portia-sdk-python==0.7.0
posthog==6.6.0
prompt-toolkit==3.0.41
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.16.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
shellingham==1.5.4
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.43
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
tenacity==9.1.2
testcontainers==4.12.0
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.66.1
traitlets==5.13.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.0.7
uvicorn==0.35.0
wcwidth==0.2.10
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
